+++ app/src/test/kotlin/no/nav/tms/varsel/authority/database/LocalPostgresDatabase.kt
@@ -3,7 +3,6 @@ package no.nav.tms.varsel.authority.database
 import com.zaxxer.hikari.HikariDataSource
 import kotliquery.queryOf
 import no.nav.tms.varsel.authority.common.Database
-import no.nav.tms.varsel.authority.common.toJsonb
 import org.flywaydb.core.Flyway
 import org.testcontainers.containers.PostgreSQLContainer
 
@@ -53,18 +52,6 @@ class LocalPostgresDatabase private constructor() : Database {
             .let { assert(it.migrationsExecuted == expectedMigrations) }
     }
 
-    fun insertLegacyArkiverteVarsler(ident: String, vararg varsler: ArkiverteDbVarsel) {
-        varsler.forEach {
-            insertArkivertVarsel(ident, it.id, it.legacyJsonFormat())
-        }
-    }
-    fun insertCurrentArkiverteVarsler(ident: String, vararg varsler: ArkiverteDbVarsel) {
-        varsler.forEach {
-            insertArkivertVarsel(ident, it.id, it.currentJsonFormat())
-        }
-    }
-
-
     fun insertArkivertVarsel(ident: String, varselId: String, jsonBlob: String) {
         instance.update {
             queryOf(
@@ -81,4 +68,3 @@ class LocalPostgresDatabase private constructor() : Database {
         }
     }
 }
-
